Support typescript output on xstate transpiler ⚠️
* core Builder now support custom option key

* xstate transpiler also support javascript output

⚠️ only applied on `scrap code` subcommand
